armor, crafting recipes, fuel rework, massive combining of classes
huge update to the code. started to implement armor, crafting recipes, and a more advanced system to init items. I combined a lot of classes together, and I am working on combining more. as for the moment, the crusher is still broken, but I am still working on fixing it.